// types/next-auth.d.ts
import { DefaultSession } from "next-auth";

export type UserRole = "admin" | "walas" | "bk" | "siswa";

declare module "next-auth" {
  interface User {
    role: UserRole;
  }

  interface Session {
    user: {
      id: string;
      role: UserRole;
    } & DefaultSession["user"];
  }
}

declare module "next-auth/jwt" {
  interface JWT {
    id: string;
    role: UserRole;
  }
}